move yaml loader into core module and remove yaml_logic module
