+++ src/nominatim_data_analyser/core/core.py
@@ -1,7 +1,7 @@
 import logging
 from pathlib import Path
 
-from .yaml_logic.yaml_loader import load_yaml_rule
+from .yaml_loader import load_yaml_rule
 from .assembler.pipeline_assembler import PipelineAssembler
 from ..timer import Timer
 from ..config import load_config
